move kept legacy screens
