+++ src/screens/WalletConnect/WalletConnectConnectorRequest.js
