+++ src/screens/WalletConnect/WalletConnectPinConfirm.js
